optimized out fp64 instructions

+++ src/include/vec3.cuh
@@ -68,15 +68,15 @@ __host__ __device__ inline Vec3 operator*(const Vec3 &u, const Vec3 &v) {
   return Vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-__host__ __device__ inline Vec3 operator*(double t, const Vec3 &v) {
+__host__ __device__ inline Vec3 operator*(float t, const Vec3 &v) {
   return Vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-__host__ __device__ inline Vec3 operator*(const Vec3 &v, double t) {
+__host__ __device__ inline Vec3 operator*(const Vec3 &v, float t) {
   return t * v;
 }
 
-__host__ __device__ inline Vec3 operator/(const Vec3 &v, double t) {
+__host__ __device__ inline Vec3 operator/(const Vec3 &v, float t) {
   return (1 / t) * v;
 }
